异常处理 处理用户名重复的异常 添加、修改成员信息时小组id设置异常 删除小组时校验是否有成员

+++ system/src/com/alana/javaweb/service/GroupService.java
@@ -1,12 +1,14 @@
 package com.alana.javaweb.service;
 
 import com.alana.javaweb.dao.GroupDao;
+import com.alana.javaweb.exception.BusinessException;
 import com.alana.javaweb.model.Group;
 
 import java.util.List;
 
 public class GroupService {
     private GroupDao groupDao = new GroupDao();
+
 
     public boolean addGroup(Group group) {
         return groupDao.insert(group) > 0;
@@ -18,6 +20,10 @@ public class GroupService {
     }
 
     public boolean deleteGroup(int groupId) {
+        MemberService memberService = new MemberService();
+        if(memberService.checkMemberExists(groupId)){
+            throw new BusinessException("小组内还有成员");
+        }
         return groupDao.delete(groupId) > 0;
     }
 
@@ -31,6 +37,10 @@ public class GroupService {
     }
     public List<Group> listGroups(){
         return groupDao.selectAll();
+    }
+
+    public boolean checkGroupExists(int groupId) {
+        return (groupDao.selectById(groupId) != null);  //不等于null→存在→直值
     }
 }
